requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
numpy==1.24.3
pandas==2.0.1
textblob==0.15.3
vaderSentiment==3.3.2
